+++ create-echoed/src/bin.ts
@@ -1,8 +1,8 @@
 #!/usr/bin/env node
 
 import { AnsiGreen, AnsiRed, AnsiReset } from "@/ansi";
-import { Creator } from "@/creator";
-import { program } from "commander";
+import { Creator, TEMPLATE, TEMPLATES } from "@/creator";
+import { Option, program } from "@commander-js/extra-typings";
 import fs from "fs";
 import path from "path";
 
@@ -11,13 +11,21 @@ const ECHOED_ROOT_DIR = path.resolve(__dirname);
 program.name("echoed");
 
 program
-  .command("jest", { isDefault: true })
+  .addOption(
+    new Option("-t, --template <template-name>")
+      .choices(TEMPLATES)
+      .default<TEMPLATE>("jest"),
+  )
   .description("create sample files for Echoed")
   .action(runCreateSample);
 
 program.parse();
 
-async function runCreateSample(): Promise<void> {
+async function runCreateSample({
+  template,
+}: {
+  template: TEMPLATE;
+}): Promise<void> {
   const cwd = process.cwd();
   if (fs.existsSync(cwd)) {
     if (fs.readdirSync(cwd).length > 0) {
@@ -32,7 +40,7 @@ async function runCreateSample(): Promise<void> {
   }
 
   const creator = new Creator(ECHOED_ROOT_DIR, cwd);
-  await creator.createJestExample();
+  await creator.create(template);
   console.log(`Finish initializing Echoed.
 
 The initialization also creates example tests to demonstrate Echoed's behavior.